Index & Validation
In this commit the following is contain:

Validation logic to make sure the proper files are submitted to the proper locations; i.e., a sample file cannot be uploaded as a baseline, as well as (vice versa) a baseline file cannot be uploaded as a sample file (and will even point out which file it is that causes the error).

Additionally, fixed the error message to be descriptive for the user and so there is no longer a disconnect between the javascript obtaining the error message and the backend.